+++ boofuzz/mutation_context.py
@@ -1,7 +1,17 @@
 import attr
+import collections
 
 from .mutation import Mutation
 from .protocol_session import ProtocolSession
+
+
+def mutations_list_to_dict(mutations_list_or_dict):
+    if isinstance(mutations_list_or_dict, dict):
+        return mutations_list_or_dict
+    elif isinstance(mutations_list_or_dict, collections.Iterable):
+        return {mutation.qualified_name: mutation for mutation in mutations_list_or_dict}
+    else:
+        raise ValueError("Cannot initialize a MutationContext with mutations {0}".format(mutations_list_or_dict))
 
 
 @attr.s
@@ -18,5 +28,6 @@ class MutationContext(object):
     ProtocolSession does not necessarily have a MutationContext.
     """
 
-    mutation = attr.ib(type=Mutation)
+    mutations = attr.ib(factory=dict, converter=mutations_list_to_dict)  # maps qualified names to a Mutation
+    message_path = attr.ib(factory=list)
     protocol_session = attr.ib(type=ProtocolSession, default=None)