+++ CHANGELOG.rst
@@ -1,9 +1,19 @@
 Changelog
 =========
 
+Upcoming
+--------
 Features
 ^^^^^^^^
+- Fuzzing CLI -- Use main_helper() to use boofuzz's generic fuzzing CLI with your script.
+- Combinatorial fuzzing -- now fuzzes multiple mutations at once by default.
+- Added `Simple` primitive that uses only the specified values for fuzzing.
+- Fixed two memory leaks in the fuzz logger.
+- Test cases can now be specified and re-run by name.
+- Implemented visual request-graph rendering functions for Session
+- Added to web UIL: runtime, exec speed, current test case name.
 - Added simple custom checksum and example usage.
+
 Fixes
 ^^^^^
 - Clarified Checksum() data for custom checksum function.